add named routing for genebanks
